Initialize to the current status
The basic script, as it has been used so far. Its use involves manually
changing the parameters of the various ffmpeg commands, adding new ones,
and removing unused ones. For example, in the current status it is set
to create the stereo soundtrack by extracting the surround sound from a
file and converting it using ffmpeg filters. Other times there is an
existing stereo track, and only the volume of that needs adusting. In
other cases there is both a stereo, and a surround sound track and both
need to be adjusted. Each use requires rewriting the script to do what
is needed and removing functionality from previous uses that don't fit
the current use.

The whole point of this project is to create a script (or collection of
scripts) that can choose which functionality is needed for the current
source files. Some sort of config file will be needed to hold the unique
settings for each use. Another point will be creating a method to only
re-run the commands that are needed as adjustments to the audio and
video are needed.

I have used the current version long enough to have established the
common setting, which usually get the results I want. There are cases,
however, where an inspection of the final result shows that different
settings are needed for either the volume or some aspect of the video.
Currently this means editing the script for the new values, and running
the entire building process from scratch. A simple increase in the
volume, therefore, also means re-encoding the video, re-encoding all of
the audio, rather than just the new stereo track, and recreating the
subtitle files. The same, complete, process has to be run to correct a
mere typo I find in the subtitles.

I want to make it so that the component parts are extracted, adjusted,
and encoded individually, and the process is halted before the final
remuxing. That allows me to inspect each component on its own, using
SMPlayer to view the proposed remuxing before doing it. Any component
which needs adjustment can be re-encoded, while retaining the encoded
version of all the other components. Once I am satisfied with the final
inspection, one last command can create the final product by remuxing
the created, and inspected, components.

I want to create the whole using a modular system. Doing so will allow
for future enhancements, replacement of the tools used, etc., without
requiring a complete rewrite of the whole system.

One final objective is to make the system slightly more user-friendly by
having it, at the least, provide some clue as to the progress being made
in creating the component parts. As it runs now, the only clues are the
statistics for the current ffmpeg process:
> frame=219180 fps=176 q=27.0 Lsize= 2216285kB time=02:32:21.63 bitrate=1986.1kbits/s speed=7.33x
is not very helpful unless I already know what's been done and what is
yet to come. That could be the splitting of the surround sound into
channels, extracting stereo from the surround sound, adjusting the gain
on the stereo soundtrack, extracting the hearing-impaired subtitles, or
even the re-encoding of the video. More output for an informed user is
something I've been wanting, but haven't actually added. Making this a
"project" will help me to 'just do it' rather than wishing for it to be
done.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>