Split the filename usage into working and final versions
Split the filename variables to handle one name, probably shorter, for
the files created in the process (the working files), and another for
the name used when all final product files are created.

Also increased the shell safety by switching to single quotes wherever a
filename or user-generated string is used in something the shell will
see.

Switched the majority of other double quotes to single quotes as a
matter of cosmetics. Double quotes are kept where variable expansion is
needed or where they help with understanding something internally.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>