Factor out the surround sound to stereo conversion
Continuing the process of converting the monolithic, hand edited script,
the process of extracting the surround sound and converting it to stereo
is in the hts-stereo command. Nearly identical to the stereo command in
its functionality, and using the same enhancement filter it has the same
effects, and limitations, as the stereo command.

Added new command to the usage file.

Corrected some typing errors in the stereo command.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>