Factor out the scanning of source files
Scanning of the source candidate files, done in the init script has been
removed to its own script, explore. This allows the individual files, or
new files added to the directory, to be scanned later by the user while
not forcing a complete rescan of everything by calling init.

Added new command to the usage file.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>