Factor out surround sound processing
Added hts-split, hts-normalize, and hts-merge for processing the
surround sound streams. The choice to separate the processing into 3
steps was based on use case scenarios encountered and processing times.

The process of finding the 'right' values for normalizing the soundtrack
can be a cycle of encoding and testing. The default values often come
close, but still need some adjustments, sometimes requiring multiple
passes to find the needed settings. That process is built into the
hts-normalize command. The hts-split command only splits the channels
out of the stream, and the hts-merge command recombines the channels to
create the stream. The time taken to split or merge the channels can be
rather long, and only needs to be done once. The normalizing process is
the only step that needs repeating, and only uses about 10% of the
processing time out of the entire process. Repeating the other 90% for
each iteration of trials seems like a waste of time.

The creation of DTS formated surround sound has been completely removed
from the normal work flow, with the option of creating DTS files given
with the dts-build command. Reading forums suggests that the DTS format
is far inferior to the AC3 format in compression and quality, while it
is also signicantly slower, about 6 times slower, to encode. In the
event any of the hts-* commands are called out of order, or if the
working files created by earlier steps are missing for some reason, the
called command will force the preceding command to be run. Hence, if the
initial command run is hts-merge, it will not find the needed file and
force hts-normalize to run. In turn, hts-normalize will find the files
it needs missing, and force hts-split to run. Once the needed files
exist, the command that called the preceding command will run. A single
call to hts-merge will result in the eventual creation of all the files
in the process, and is a useful method of operating when it is known, or
suspected, that the default normalizing setting will be "good enough."

Earlier commands have been cleaned up some, attempting to keep some
uniformity between the layout of the command files. Typos have been
corected as they are found. The usage file has been expanded to include
the new commands, and reformated to 80 column screens.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>