Factor out video extraction
Create the command for extracting, and modifying, the video from source
file and create the version of the video that will become the final
product. This is the part to make one version of the video. There is a
seperate command that will act as the launcher for this and control how
many videos are made, and with what parameters.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>