Add Community-building files - DOC
Include the GitHub-recommended file, in this case Developer Certificate
of Origin over Contributor License Agreement.

A CLA covers too much territory, and would be the largest file in the
repo as well. The DOC used here is the same one used by the Linux coders
at the Linux Foundation. If it has worked for this long, for the number
of contributors involved, and for the scale of the project, surely it is
good enough for anything I will ever create.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>