Add ability to handle multiple subtitle versions
Need the ability for having multiple version of the subtitles associated
with a video. For example, normal subtitles, hearing-impaired subtitles,
and subtitles with only foreign parts translated.

Requires updating the config file, reworking the extracting file, and
adding a launch control file. The system will be very similar to that in
use for control of video file extraction.

Also removed the 'ghost file' extract_video, which was renamed to
extract-video, git kept the old and made a new copy. (Operator error.)

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>