Convert ffmpeg execution to subshell
Using a intermediate command file was cluttering up the directory too
much. Still want to include the log file, clutter or not, as they can be
helpful in deciding how to adjust the next run of the command when the
results are not what is desired.

Add wave format output from stereo command to make visual comparison of
the wave form faster using Audacity. (Wave files load faster than AAC
or AC3 file do.)

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>