Use a common source file
Updated the config file to place the commonly changed items at the end,
making them easier to find and update. Also changed internal parameters
so that it is possible to "run" the system without any changes to the
config file. (See below about single source file)

Updated the extraction commands to check for the setting in their source
file name, and use the video source file if there is not one set for the
specific purpose, such as audio or subtitles.

The init command now looks for, and scans, a file named 'movie', without
any extension, as the first target. The same name is used in the config
file as the video source. FFMPEG does not need the file extension as a
way of knowing what kind of file is used as the input. By renaming the
source file, or creating a symlink to it, using 'movie' as the full file
name it is possible to not need to change anything in the config file
for MKV and MP4 files, and minimal changes for AVI files.

The change from multi-source to single source also required that the
process of finding the valid range of subtitle setting groups use some
other "key" than the source name. Switching to the stream number, which
is still required, solved that issue.

Other minor typos were cleaned up as well.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>