Add the ability to include dvd_subtitle format subtitles
New command sutitles-dvd to process subtitles in the dvd_subtitle format
commonly found in MKV files. They are "passed through" to the resulting
video files without being converted to text-based.

Inserted the  extra logic into subtitles and build-video commands.

Convert surround sound layout to work based on a selection in the group
of values at the end of the file, and make the different options have
their own data with the command files concerned implementing logic to
select the correct data based on the single setting at the end of the
config file.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>