Update copyright notices
Copyright notices in files recently edited have been updated to reflect
the new year.

The convention for generated file names has been altered to use hypen in
place of underscore for the separation of name parts. My personal choice
is to use underscores to replace spaces in the names of files, and the
hyphen to separate elements. Most video have a multi-word name and
seeing the underscore character repeated across the entire name was not
helpful when trying to find the year, resolution or codec notations in
the file/folder names. Using the hyphen helps keep the elements separate
from the video name, which is split with the underscore.

Drop the -pix_fmt from the video filter graphs. Added lines to include
it as an option when downsapling from 10-bit to 8-bit. It was added for
doing such downsampling, and seemed harmless to leave in for standrad
use. Turns out that it has a negative impact when processing videos that
are already 8-bit, and had to be removed, but not forgotten.

Add a changelog file, in preparation for a new relase.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>