Adjust normalization and stereo mixing parameters
Change the compression factor for normalization from 30, the minimum,
to 20. Dynamic compression was set to the min for a few videos, but it
seems to be doing more harm than good for most other videos. I'm trying
to find the best default setting to include in the config file. They are
intended to be adjusted as needed anyway, but sensible defaults as a
starting point seems like a good idea.

Change the channel layout for mixing stereo from 5.1 surround sound
sources. Using the `FFPEG` documentation as a guide, I found:
[Audio channel layouts][https://ffmpeg.org/doxygen/3.0/group__channel__mask__c.html]
I keeping with that, which is suggestive to someone not "in the know"
about audio codecs and internal structures of the files, the layout
used _might_ have been wrong.

Minor adjustment to the channel ratios in the downmix to strero.

Reworded some of the internal documentation for `hts-stereo` command.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>