Add 7.1 surround sound and parallel processing
Add the ability to process 7.1 surround sound streams. All the `hts-*`
commands utilize a standard method of pulling the surround sound data
from the config file. Possible consideration to make it into a shared
function somewhere.

The rework of the channel processing for 7.1 ability also should make it
a simple process to add other sound layouts as needed. (Even easier if
a shared function can be created.)

Added the ability to process commands in parallel.  `FFMPEG` reportedly
is good at optimizing thread usage. My running of commands with a select
few `threads` options, including 0 (auto?) never managed to utilize my
multi-core CPU to any significant level when processing audio files. To
make better use of my system I started doing commands in parallel. To
that end I liberally copied a section of code from Stack Overflow:
[How do you run multiple programs in parallel from a bash script?](https://stackoverflow.com/a/42098494/7412956)

Gains from parallelism are margin, yet a net positive, so the change
remains. Rare conditions may result in a performance decrease, but I
believe the rare events are insufficient to remove the concept. The
process most benefited by parallelism is the multi-channel normalize
process. Video of 2h 57m with 5.1 sound using 6 (or more) parallel
processes, relative to single processes, uses ~37% less time.

The output from the `FFMPEG` commands and the updates from the process
of handling parallels often intermingles. Though it is "ugly" to see, my
purpose for the `FFMPEG` output is to see that processing _is_ happening
rather than logging or documenting the process. The log files do a good
job of that. Keeping the output "clean" left me wondering if commands
were running, and how far into the encoding they were. For the nonce, I
esteem ugly feedback more than clean silence.

Reverted the change for channel layouts. Expirements proved that the the
original version was correct, and that the sequence used by FFMPEG is
somehow internal usage, but not indicative of the arrangement within the
audio streams.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>