Factor out group processing for video and subtitles
The determination of valid group numbers for the video and subtitle
extraction process was determined in a common manner, but was repeated
in each file. The same determination in the video building command was
done in a slightly different manner. In all cases, and for future use,
the determination has been moved into a common function and the work
of processing the items in the group has been converted into a function
which the common function can handle in an abstract fashion.

Now there is less work to make a change in the determination process,
and less work to add groups to other items, such as audio streams.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>