Add language parameter to subtitle creation
When extracting subtitles there often is not a language assigned to them
in the source file. The processing of the subtitles may also remove it
when there is one. In either case, the ability to set that into the
created files was lacking from the process.

The main program sets the default, or fall back, language for everything
to 'und', the code for undefined. The config file is then read, which is
able to override that with a 'video_lang' setting. In all cases where a
language is set, if there is not one set for the file, or stream, being
created, the default will be used.

Moved more common work from individual extraction commands into the main
dispatcher. Added parallelism to subtitle extraction. Seldom do videos
have more than one, if that, subtitles, so there is likely little
benefit to parallel processing. Set up for it nevertheless.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>