Enable processing of multiple audio streams
Converted the multiple video and subtitle processing algorithms for use
in processing the audio streams to enable the processing of more than
one audio stream.

Requires that the config file be modified to use suffix digits on the
audio settings in the same manner as the subtitles and video settings.

The conversion of the DTS stream creation remains to be done.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>