Rework final product stream selection logic
With the potential for a plethora of generated sound files, the choice
of which ones to include becomes wider than my personal goals. What I
want to build are videos with a sinlge enhanced stereo stream. The use
of a single stream keeps the file size low, the enhanced sound helps my
ears (part of the genisis of the project), and my system only has stereo
sound available.

For usefulness to others, the system needs to be able to include any of
the generated streams, including multiple versions of them, into the
final video file. In recognition of that, the audio section in the
config file has been expanded, and reworked. The commands to create the
streams have been updated to match.

All the audio processing commands and sub-commands have been affected by
the change. There should be very limited changes needed to embed the
creation process into the final video building command. The functions
involved in the use of some, as yet unknown, codec for audio streams has
been considered, and some groundwork for that eventuality has been done.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>