Remove forced inclusion of audio streams
The inclusion of at least one version of _every_ audio stream is not a
good choice. The real objective should be that there is _at least one_
audio stream included in the final product. The logic for that should be
in the process for building that file rather than forcing inclusion of
all streams.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>