Change output names in audio stream processing
Alter the names for generated audio streams so that there was a common
convention between versions. They are group better in the directory list
and are easier to "pull in" when doing the rebuild of the final product.

-mono-?.wav files are from the processing of individual channels in the
surround sound streams. The extraction creates files with channel names
and the normalizing process prepends an 'a' to the channel names.

-remix.* and -remix-norm.* files are surround sound streams, in original
and normalized volumes. The extensions, selectable in the config file,
are .ac3, .aac, and .dts

-stereo.* and -stereo-norm.* files are stereo streams, in original and
normalized volumes. Then extensions, selectable in the config file, are
.ac3, and .aac. The name no longer indicates if the source was a stereo
stream or a surround sound stream. If the audio_label_# setting is used
with unique names for each source, as it should be, the generated files
will still be identified by the user, and will not clobber each other as
they are made.

The uniqueness of the ?_label settings may become even more significant
as I attempt to increase the level of parallel processing within the
system as a whole.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>