Refactor helper functions
The helper functions have been expanded and refactored to take on more
of the work which was being repeated in multiple files. Almost all the
work done by the helper scripts is loading specific settings into the
general settings used by other processes. All the other files will need
to be updated to implement the new versions.

The functionality of preparing the environment for processing each of
the component elements has been removed to distinct functions rather
than being a large block in the process_group function. Each of the
component types now has their own prefix in the general settings so that
the processing of one can be chained to another, or paralleled to others
without clobbering the settings.

The settings work done by the multi_channel_setup function is done along
with all other settings work, while the logic has been removed to the
places where it is needed rather than being communicated with error code
values.

Laying the groundwork for the conversion from using the source command
to using function calls is included in the update.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>