Change from sourced files to functions - remix
Begin the process of converting from calling sourced script files to the
use of function calls.

The primary reason being that as function calls every thing will have
the same parent process, and the parallel processing fan be shared
across all the commands rather than being limited to use within each of
the commands.

During the process of conversion, the files will be "cleaned" as well.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>