Gather functions into related files
With all commands now implemented as functions rather than as sourced
and executed scripts, there is no need to have each command in its own
file.

All the sound work has been gathered into the audio file. The one
dispatcher, process_group, handles all the "decision" work for which of
the group settings to use. That functionality was more or less built in
to each of the command file. That function is in the common functions
file. The subtitle and video extraction each have their own file of
functions and the recombining function, remux, has its own file of
functions.

The 'usage', or help, function has been rewritten as well. Now a simple
help command lists available commands with short, one-liner, versions of
their purpose. The help command takes another command as an argument and
supplies a longer description of the command and its use.

Added help topics for the target_num and force arguments, the dynaudnorm
audio filter and links to FFMPEG site for the video filter docs.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>