Change default timestamp
The timestamp used in the metadata for the files is set to the current
time when the script is launched. It is still in the fixvid.conf file,
so the user is able to put any value there they desire. Such as for when
the video was encoded previously, and a new soundtrack, or subtitle
stream is being added to it.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>