Add finalize ability to titles command set
The finalize ability copies the created files marked as `_save_` in the
config file into a subdirectory, ready for exporting. Adding the ability
to all command sets will be a step-by-step process

At the same time, the help and usage functionality and data will move
into a `help` file and the data will be updated to track the changes in
the scripted functions themselves.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>