Add finalize ability to video command set
The finalize ability copies the created files marked as `_save_` in the
config file into a subdirectory, ready for exporting. Adding the ability
to all command sets will be a step-by-step process

At the same time, the help and usage functionality and data will move
into a `help` file and the data will be updated to track the changes in
the scripted functions themselves.

The remux command is really just part of the video command set, so that
file is merged into the video scripts. The general functions from remux
are added to the functions file instead, as that is where they belong.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>