Add option for a minimalistic version of fixvid.conf
Created a new version of the config file without comments and with only
one group of settings for each type of stream.

Added a parameter to the init command that will generate that version in
place of the full version.

Created a single-file version of the program. Saving around 13% of the
size of the main files.

Did some cleanup of the files, trying to give them all the same "look"
in terms of how the code is presented. Doing so also resulted in a few
minor modifications to the code itself, for clarity or for "convention."

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>